make physdim magnitude int, not size_t (do not need to use long on 64bit archs)

+++ phys.h
@@ -5,7 +5,7 @@
 
 struct physdim {
 	char			*pd_name;
-	size_t			 pd_mag;
+	int			 pd_mag;
 	double			 pd_space;
 	struct physdim		*pd_contains;
 	struct physdim		*pd_containedby;
